Moved the process startup functions into their respective modules
